Finished the configuration of Resolver Dependency Injection
