+++ timetable/Presentation/Views/EventListView.swift
@@ -14,8 +14,6 @@ class EventListView: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
     }
 
 }
