+++ timetable/SceneDelegate.swift
@@ -6,12 +6,13 @@
 //
 
 import UIKit
+import Resolver
 
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 
     var window: UIWindow?
     
-    private lazy var navigationController = UINavigationController(rootViewController: EventView())
+    private lazy var navigationController = UINavigationController(rootViewController: Resolver.resolve(EventListView.self))
 
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
         guard let scene = (scene as? UIWindowScene) else { return }
